* check-point 1 - at this point, i canceled the reading of RootCA bytes in the sidecar for orderer connection, we need to understand how to create easily config-block with the root-CA we currently use for-each test.
In addition, i need to create an interface for the config-block converters that each of them will have.

each config that is not config parameter need to implement the converter function.

I need to create a father struct that contains the mutual fields in the Config and its twin, ConfigParameter.

I need to understand if I need to change something in the adapters that now have been affected by the change.
And, re-read the connection-manager package and look if there are any mistakes regarding the 'trueness' of the implementation after changes.

* This commit should be running all the tests correctly but the linter.

Signed-off-by: Dean Amar <Dean.Amar@ibm.com>